New Development: no             JIRA Issue: no       Ready to Release: yes      Interface Changes: none What Interface Changed: none          Test Programs: none   Put in Release Notes: no
           Description: changes required for building casacore, code, and gcwrap on OSX 10.10 with Xcode 6



git-svn-id: https://svn.cv.nrao.edu/svn/casa/trunk@32632 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ code/synthesis/MeasurementComponents/StandardVisCal.cc
@@ -1048,7 +1048,7 @@ void BJones::calcWtScale() {
 
     // the weight scale factor is just the square of the 
     //   freq-dep normalization
-    cWSi.array()*=pow(calcPowerNorm(A.array(),!Aok.array()),2);
+    cWSi.array()*=(float)pow(calcPowerNorm(A.array(),!Aok.array()),2);
 
     A.next();
     Aok.next();
